==> 1_web_scraping_using_stocktwits.ipynb <==
# importing libraries for data manupulating 
import pandas as pd 

# Importing libraries for web_scraping 
import requests
import json
import datetime


## Web Scraping from stocktwits 

def stocktwits_scrap(n,ticker,base=179808000):
    id=[]
    created_at=[]
    body=[]
    for i in range(n):
        try:
            url = "https://api.stocktwits.com/api/2/streams/symbol/"+ticker+".json?max="+str(base+5000)+"&since="+str(base)+"&limit=30"

            base=base+5001

            response = requests.request("GET", url, headers={}, data ={})
            response=response.json()
            messages=response['messages']
            iter=range(len(messages))
            for idx in iter:
                id.append(messages[idx]['id'])
                created_at.append(messages[idx]['created_at'])
                body.append(messages[idx]['body'])

                if (len(id) % 100) == 0:
                    print('Done Scrape ',len(id),' messages.')
        except:
            print('Warning Messages: \n ID Out of Range or Too Many Requests within 1 Hour')
            break
    try:
        df_stocktwits=pd.DataFrame(id,columns=['id'])
        df_stocktwits['date_time']=created_at
        df_stocktwits['date_time']=pd.to_datetime(df_stocktwits['date_time'])
        df_stocktwits['tweet']=body
        df_stocktwits=df_stocktwits.sort_values(by='date_time')
        df_stocktwits=df_stocktwits.reset_index(drop=True)
        print('Done Scrape ',len(id),' messages.')
        print('Last ID: ',df_stocktwits.tail(1)['id'].values[0])
        return(df_stocktwits,df_stocktwits.tail(1)['id'].values[0])
    except:
        print('Error')
        return None,last_id


def main():
        # Technology : AAPL (Apple)
        # Semiconductors: NVDA (NVIDIA Corporation)
        # Financial : MA (Master Card)
        # Consumer defensive : PG (Procter & Gamble Co)
        # Healthcare : PFE (Pfizer inc.)
        
        n = 5 
        
        d = {"technology":"AAPL","semiconductors":"NVDA","financial":"MA",
             "consumer_defensive":"PG","healthcare":"PFE"}
        for ticker in d:
            df_stocktwits, last_id = stocktwits_scrap(n,d[ticker])

            df_stocktwits.to_csv("df_stocktwits_"+ticker+".csv")
    
if __name__ == '__main__':
        main()

df = pd.read_csv("df_stocktwits_technology.csv", usecols=['id','date_time','tweet'])
df[:7]

# Industries considered in paper         
#finance : 
# Manufacturing: 
# Mining: 
# Energy : 
# Transportation:
