==> 3_Dataset merging.ipynb <==
# Importing libraries for working with datasets
import numpy as np
import pandas as pd 

# importing libraries for data visualisation 
import matplotlib.pyplot as plt
import seaborn as sns 

# Importing libraries for web sracping 
import requests
import json
from bs4 import BeautifulSoup
import datetime
import time 
import csv


import gensim

### import the datasets

df_aapl_1 = pd.read_csv("datasets/df_stocktwits_AAPL-1.csv")
df_aapl_2 = pd.read_csv("datasets/df_stocktwits_AAPL-2.csv")
df_aapl_3 = pd.read_csv("datasets/df_stocktwits_AAPL-3.csv")
df_aapl_4 = pd.read_csv("datasets/df_stocktwits_AAPL-4.csv")
df_aapl_5 = pd.read_csv("datasets/df_stocktwits_AAPL-5.csv")
df_aapl_6 = pd.read_csv("datasets/df_stocktwits_AAPL-6.csv")
df_aapl_7 = pd.read_csv("datasets/df_stocktwits_AAPL-7.csv")
df_aapl_8 = pd.read_csv("datasets/df_stocktwits_AAPL-8.csv")
df_aapl_9 = pd.read_csv("datasets/df_stocktwits_AAPL-9.csv")
df_aapl_10 = pd.read_csv("datasets/df_stocktwits_AAPL-10.csv")




df_googl_1 = pd.read_csv("datasets/df_stocktwits_GOOG-1.csv")
df_googl_2 = pd.read_csv("datasets/df_stocktwits_GOOG-2.csv")
df_googl_3 = pd.read_csv("datasets/df_stocktwits_GOOG-last.csv")



df_ma_1 = pd.read_csv("datasets/df_stocktwits_MA-1.csv")
df_ma_2 = pd.read_csv("datasets/df_stocktwits_MA-2.csv")
df_ma_3 = pd.read_csv("datasets/df_stocktwits_MA-3.csv")

df_amzn_1 = pd.read_csv("datasets/df_stocktwits_AMZN-1.csv")
df_amzn_2 = pd.read_csv("datasets/df_stocktwits_AMZN-2.csv")
df_amzn_3 = pd.read_csv("datasets/df_stocktwits_AMZN-3.csv")




df_jnj_1 = pd.read_csv("datasets/df_stocktwits_JNJ-1.csv")
df_jnj_2 = pd.read_csv("datasets/df_stocktwits_JNJ-2.csv")
df_jnj_3 = pd.read_csv("datasets/df_stocktwits_JNJ-3.csv")
df_jnj_4 = pd.read_csv("datasets/df_stocktwits_JNJ-4.csv")
df_jnj_5 = pd.read_csv("datasets/df_stocktwits_JNJ-5.csv")


#df_ma_1

# concatinating the datasets 
df_aapl = pd.concat([df_aapl_1,df_aapl_2,df_aapl_3,df_aapl_4,df_aapl_5,df_aapl_6,df_aapl_7,df_aapl_8,df_aapl_9,df_aapl_10],axis=0)


df_googl= pd.concat([df_googl_1,df_googl_2,df_googl_3],axis=0)

df_amzn = pd.concat([df_amzn_1,df_amzn_2,df_amzn_3], axis=0)

df_jnj = pd.concat([df_jnj_1,df_jnj_2,df_jnj_3, df_jnj_4, df_jnj_5], axis=0)

df_ma = pd.concat([df_ma_1,df_ma_2,df_ma_3],axis=0)


df_aapl

def clean_columns(df):
    
     # remove extra column 
    df.drop(['Unnamed: 0','id'], axis=1, inplace = True)  
    
    # removing the time from date_time column, keeping only date
    df['date_time'] = pd.to_datetime(df['date_time'])
    df['date_time'] = df['date_time'].dt.date
    return df


df_list = [df_aapl, df_googl,df_ma, df_amzn, df_jnj]

for df in df_list: 
    df= clean_columns(df)

df_aapl

df_aapl.to_csv("stock_datasets/df_stocktwits_aapl.csv")
df_googl.to_csv("stock_datasets/df_stocktwits_googl.csv")
df_ma.to_csv("stock_datasets/df_stocktwits_ma.csv")
df_amzn.to_csv("stock_datasets/df_stocktwits_amzn.csv")
df_jnj.to_csv("stock_datasets/df_stocktwits_jnj.csv")

# def clean_date(df):
    
#      # remove extra column 
#     df.drop(['Unnamed: 0','id'], axis=1, inplace = True)  
    
#     # removing the time from date_time column, keeping only date
#     df['date_time'] = pd.to_datetime(df['date_time'])
#     df['date_time'] = df['date_time'].dt.date
    
#     # Removing the punctuations and http links 
    
#     df.replace(r"http\S+"," ", regex=True, inplace=True)
#     df.replace("[^a-zA-Z]"," ", regex=True, inplace=True)  

#     # Converting the tweet into lower case
#     df['tweet'] = df['tweet'].str.lower()


# df_list = [df_aapl, df_msft,df_nvda,df_googl,df_meta,df_nflx,df_ma,df_v
#           , df_brk, df_amzn, df_tsla, df_mcd, df_jnj, df_pfe, df_lly]

# for i in df_list:
#     clean_date(i)
 


# # from nltk.sentiment.vader import SentimentIntensityAnalyzer
# # import pandas as pd
# # import matplotlib.pyplot as plt

# # vader = SentimentIntensityAnalyzer()

# # f = lambda title: vader.polarity_scores(title)['compound']
# # df_mcd['compound'] = df_mcd['tweet'].apply(f)
# # df_mcd['date_time'] = pd.to_datetime(df_mcd.date_time).dt.date


# # plt.figure(figsize=(10,10))
# # mean_df = df_mcd.groupby([ 'date_time']).mean().unstack()
# # mean_df = mean_df.xs('compound')
# # mean_df.plot(kind='bar')
# # plt.show()
